# api_spec.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "FinZen API",
                 "description":  "AI-powered micro-investment platform for beginners",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/api/education/education/modules":  {
                                                           "get":  {
                                                                       "tags":  [
                                                                                    "Education",
                                                                                    "Education"
                                                                                ],
                                                                       "summary":  "Get Modules",
                                                                       "operationId":  "get_modules_api_education_education_modules_get",
                                                                       "parameters":  [
                                                                                          {
                                                                                              "name":  "authorization",
                                                                                              "in":  "header",
                                                                                              "required":  false,
                                                                                              "schema":  {
                                                                                                             "anyOf":  [
                                                                                                                           {
                                                                                                                               "type":  "string"
                                                                                                                           },
                                                                                                                           {
                                                                                                                               "type":  "null"
                                                                                                                           }
                                                                                                                       ],
                                                                                                             "title":  "Authorization"
                                                                                                         }
                                                                                          }
                                                                                      ],
                                                                       "responses":  {
                                                                                         "200":  {
                                                                                                     "description":  "Successful Response",
                                                                                                     "content":  {
                                                                                                                     "application/json":  {
                                                                                                                                              "schema":  {
                                                                                                                                                             "$ref":  "#/components/schemas/EducationModulesResponse"
                                                                                                                                                         }
                                                                                                                                          }
                                                                                                                 }
                                                                                                 },
                                                                                         "422":  {
                                                                                                     "description":  "Validation Error",
                                                                                                     "content":  {
                                                                                                                     "application/json":  {
                                                                                                                                              "schema":  {
                                                                                                                                                             "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                         }
                                                                                                                                          }
                                                                                                                 }
                                                                                                 }
                                                                                     }
                                                                   }
                                                       },
                  "/api/education/education/modules/{module_id}":  {
                                                                       "get":  {
                                                                                   "tags":  [
                                                                                                "Education",
                                                                                                "Education"
                                                                                            ],
                                                                                   "summary":  "Get Module Detail",
                                                                                   "operationId":  "get_module_detail_api_education_education_modules__module_id__get",
                                                                                   "parameters":  [
                                                                                                      {
                                                                                                          "name":  "module_id",
                                                                                                          "in":  "path",
                                                                                                          "required":  true,
                                                                                                          "schema":  {
                                                                                                                         "type":  "string",
                                                                                                                         "title":  "Module Id"
                                                                                                                     }
                                                                                                      },
                                                                                                      {
                                                                                                          "name":  "authorization",
                                                                                                          "in":  "header",
                                                                                                          "required":  false,
                                                                                                          "schema":  {
                                                                                                                         "anyOf":  [
                                                                                                                                       {
                                                                                                                                           "type":  "string"
                                                                                                                                       },
                                                                                                                                       {
                                                                                                                                           "type":  "null"
                                                                                                                                       }
                                                                                                                                   ],
                                                                                                                         "title":  "Authorization"
                                                                                                                     }
                                                                                                      }
                                                                                                  ],
                                                                                   "responses":  {
                                                                                                     "200":  {
                                                                                                                 "description":  "Successful Response",
                                                                                                                 "content":  {
                                                                                                                                 "application/json":  {
                                                                                                                                                          "schema":  {
                                                                                                                                                                         "$ref":  "#/components/schemas/ModuleDetailResponse"
                                                                                                                                                                     }
                                                                                                                                                      }
                                                                                                                             }
                                                                                                             },
                                                                                                     "422":  {
                                                                                                                 "description":  "Validation Error",
                                                                                                                 "content":  {
                                                                                                                                 "application/json":  {
                                                                                                                                                          "schema":  {
                                                                                                                                                                         "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                                     }
                                                                                                                                                      }
                                                                                                                             }
                                                                                                             }
                                                                                                 }
                                                                               }
                                                                   },
                  "/api/education/education/modules/{module_id}/quiz":  {
                                                                            "post":  {
                                                                                         "tags":  [
                                                                                                      "Education",
                                                                                                      "Education"
                                                                                                  ],
                                                                                         "summary":  "Submit Quiz",
                                                                                         "operationId":  "submit_quiz_api_education_education_modules__module_id__quiz_post",
                                                                                         "parameters":  [
                                                                                                            {
                                                                                                                "name":  "module_id",
                                                                                                                "in":  "path",
                                                                                                                "required":  true,
                                                                                                                "schema":  {
                                                                                                                               "type":  "string",
                                                                                                                               "title":  "Module Id"
                                                                                                                           }
                                                                                                            },
                                                                                                            {
                                                                                                                "name":  "authorization",
                                                                                                                "in":  "header",
                                                                                                                "required":  false,
                                                                                                                "schema":  {
                                                                                                                               "anyOf":  [
                                                                                                                                             {
                                                                                                                                                 "type":  "string"
                                                                                                                                             },
                                                                                                                                             {
                                                                                                                                                 "type":  "null"
                                                                                                                                             }
                                                                                                                                         ],
                                                                                                                               "title":  "Authorization"
                                                                                                                           }
                                                                                                            }
                                                                                                        ],
                                                                                         "requestBody":  {
                                                                                                             "required":  true,
                                                                                                             "content":  {
                                                                                                                             "application/json":  {
                                                                                                                                                      "schema":  {
                                                                                                                                                                     "$ref":  "#/components/schemas/SubmitQuizRequest"
                                                                                                                                                                 }
                                                                                                                                                  }
                                                                                                                         }
                                                                                                         },
                                                                                         "responses":  {
                                                                                                           "200":  {
                                                                                                                       "description":  "Successful Response",
                                                                                                                       "content":  {
                                                                                                                                       "application/json":  {
                                                                                                                                                                "schema":  {
                                                                                                                                                                               "$ref":  "#/components/schemas/QuizResultResponse"
                                                                                                                                                                           }
                                                                                                                                                            }
                                                                                                                                   }
                                                                                                                   },
                                                                                                           "422":  {
                                                                                                                       "description":  "Validation Error",
                                                                                                                       "content":  {
                                                                                                                                       "application/json":  {
                                                                                                                                                                "schema":  {
                                                                                                                                                                               "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                                           }
                                                                                                                                                            }
                                                                                                                                   }
                                                                                                                   }
                                                                                                       }
                                                                                     }
                                                                        },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "EducationModulesResponse":  {
                                                                        "properties":  {
                                                                                           "modules":  {
                                                                                                           "items":  {
                                                                                                                         "$ref":  "#/components/schemas/EducationModule"
                                                                                                                     },
                                                                                                           "type":  "array",
                                                                                                           "title":  "Modules"
                                                                                                       }
                                                                                       },
                                                                        "type":  "object",
                                                                        "required":  [
                                                                                         "modules"
                                                                                     ],
                                                                        "title":  "EducationModulesResponse"
                                                                    },
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
...
                                       "ModuleDetailResponse":  {
                                                                    "properties":  {
                                                                                       "id":  {
                                                                                                  "type":  "string",
                                                                                                  "title":  "Id"
                                                                                              },
                                                                                       "title":  {
                                                                                                     "type":  "string",
                                                                                                     "title":  "Title"
                                                                                                 },
                                                                                       "content":  {
                                                                                                       "type":  "string",
                                                                                                       "title":  "Content"
                                                                                                   },
                                                                                       "quiz":  {
                                                                                                    "items":  {
                                                                                                                  "type":  "object"
                                                                                                              },
                                                                                                    "type":  "array",
                                                                                                    "title":  "Quiz"
                                                                                                }
                                                                                   },
                                                                    "type":  "object",
                                                                    "required":  [
                                                                                     "id",
                                                                                     "title",
                                                                                     "content",
                                                                                     "quiz"
                                                                                 ],
                                                                    "title":  "ModuleDetailResponse"
                                                                },
...
                                       "QuizResultResponse":  {
                                                                  "properties":  {
                                                                                     "score":  {
                                                                                                   "type":  "integer",
                                                                                                   "title":  "Score"
                                                                                               },
                                                                                     "passed":  {
                                                                                                    "type":  "boolean",
                                                                                                    "title":  "Passed"
                                                                                                }
                                                                                 },
                                                                  "type":  "object",
                                                                  "required":  [
                                                                                   "score",
                                                                                   "passed"
                                                                               ],
                                                                  "title":  "QuizResultResponse"
                                                              },
                                       "SubmitQuizRequest":  {
                                                                 "properties":  {
                                                                                    "answers":  {
                                                                                                    "items":  {
                                                                                                                  "type":  "string"
                                                                                                              },
                                                                                                    "type":  "array",
                                                                                                    "title":  "Answers"
                                                                                                }
                                                                                },
                                                                 "type":  "object",
                                                                 "required":  [
                                                                                  "answers"
                                                                              ],
                                                                 "title":  "SubmitQuizRequest"
                                                             },
...
                                   }
                   }
}